chore: InterviewSetupModal.tsx 분리
